fix cycling dependencies
fyi @aeschli @aiday-mar
