+++ src/vs/editor/contrib/hover/browser/contentHoverController.ts
@@ -22,7 +22,7 @@ import { isMousePositionWithinElement } from './hoverUtils.js';
 import { ContentHoverWidgetWrapper } from './contentHoverWidgetWrapper.js';
 import './hover.css';
 import { Emitter } from '../../../../base/common/event.js';
-import { isOnColorDecorator } from '../../colorPicker/browser/hoverColorPicker/hoverColorPickerContribution.js';
+import { isOnColorDecorator } from '../../colorPicker/browser/hoverColorPicker/hoverColorPicker.js';
 
 // sticky hover widget which doesn't disappear on focus out and such
 const _sticky = false
