+++ src/vs/workbench/contrib/chat/browser/codeBlockPart.ts
@@ -8,7 +8,7 @@ import './codeBlockPart.css';
 import * as dom from '../../../../base/browser/dom.js';
 import { renderFormattedText } from '../../../../base/browser/formattedTextRenderer.js';
 import { Button } from '../../../../base/browser/ui/button/button.js';
-import { CancellationToken, CancellationTokenSource } from '../../../../base/common/cancellation.js';
+import { CancellationToken } from '../../../../base/common/cancellation.js';
 import { Codicon } from '../../../../base/common/codicons.js';
 import { Emitter, Event } from '../../../../base/common/event.js';
 import { combinedDisposable, Disposable, MutableDisposable } from '../../../../base/common/lifecycle.js';
@@ -69,8 +69,6 @@ import { ChatTreeItem } from './chat.js';
 import { IChatRendererDelegate } from './chatListRenderer.js';
 import { ChatEditorOptions } from './chatOptions.js';
 import { emptyProgressRunner, IEditorProgressService } from '../../../../platform/progress/common/progress.js';
-import { AsyncIterableObject } from '../../../../base/common/async.js';
-import { InlineChatController } from '../../inlineChat/browser/inlineChatController.js';
 
 const $ = dom.$;
 
@@ -941,37 +939,5 @@ export class DefaultChatTextEditor {
 		response.setEditApplied(item, -1);
 	}
 
-	async preview(response: IChatResponseModel | IChatResponseViewModel, item: IChatTextEditGroup) {
-		if (item.state?.applied) {
-			// already applied
-			return false;
-		}
 
-		if (!response.response.value.includes(item)) {
-			// bogous item
-			return false;
-		}
-
-		const firstEdit = item.edits[0]?.[0];
-		if (!firstEdit) {
-			return false;
-		}
-		const textEdits = AsyncIterableObject.fromArray(item.edits);
-
-		const editorToApply = await this.editorService.openCodeEditor({ resource: item.uri }, null);
-		if (editorToApply) {
-			const inlineChatController = InlineChatController.get(editorToApply);
-			if (inlineChatController) {
-				const tokenSource = new CancellationTokenSource();
-				editorToApply.revealLineInCenterIfOutsideViewport(firstEdit.range.startLineNumber);
-				const promise = inlineChatController.reviewEdits(firstEdit.range, textEdits, tokenSource.token);
-				response.setEditApplied(item, 1);
-				promise.finally(() => {
-					tokenSource.dispose();
-				});
-				return true;
-			}
-		}
-		return false;
-	}
 }
